pyproject.toml
[project]
name = "MediaPlayer_by_text_display"
version ="2.0.0"
dependencies = [
    "ffmpeg-progress-yield>=1.0.2",
    "mcrcon>=0.7.0",
    "pathlib>=1.0.1",
    "pillow>=11.3.0",
    "tqdm>=4.67.1",
]
